M0: lift ALP encode/decode bodies into o11y-codec-rt-alp
Final M0 step. Moves all remaining ALP / Delta-ALP pure logic from
o11ytsdb/rust/ into the workspace crate. Reduces o11ytsdb/rust/'s
ALP-related modules to extern "C" shims.

Approach: replace process-global `static mut` scratch buffers
(`ALP_INTS`, `ALP_EXC`, `ALP_EXC_U64`, `DELTA_VALS`) with stack-local
arrays inside each function. Worst-case residency ~50 KB across
nested calls (delta_alp_encode → alp_encode), well within the wasm32
1 MB default stack. The `ALP_EXC_MODE` config flag stays in
o11ytsdb's `alloc.rs`; its value is read at call sites and passed as
a `delta_for_exceptions: bool` parameter to `alp_encode` /
`delta_alp_encode`.

Workspace alp crate gains:
- `alp_encode(vals, out, delta_for_exc) -> usize`
- `alp_decode_regular(input, out) -> usize`
- `alp_find_exponent(vals) -> u8`
- `decode_exceptions(...)` (with delta-FoR + plain FoR sub-paths)
- `delta_alp_encode(vals, out, delta_for_exc) -> usize`
- `delta_alp_decode_range(input, lo, hi, out) -> usize`
- `decode_values_alp(input, out) -> usize` — top-level dispatcher
- `decode_values_alp_range(input, lo, hi, out) -> usize`
- `is_delta_alp_candidate(vals, reset_count) -> bool`
- `DELTA_ALP_TAG = 0xDA` constant

o11ytsdb/rust/src/ changes:
- alp.rs: deleted (pure logic moved).
- alp_exc.rs: deleted (pure logic moved).
- delta_alp.rs: rewritten as 4 extern "C" shims (encodeValuesALP,
  decodeValuesALP, decodeValuesALPRange, encodeValuesALPWithStats).
  Reads ALP_EXC_MODE and passes the bool through.
- batch.rs: imports from o11y_codec_rt_alp; reads ALP_EXC_MODE in a
  local helper.
- range_decode.rs: stack-allocates its timestamp scratch instead of
  borrowing the (now-gone) ALP_INTS static.
- verification.rs: imports is_delta_alp_candidate from workspace.
- lib.rs: drops `mod alp;` and `mod alp_exc;`.

Verification:
- 19 codec-rt-core + 26 codec-rt-xor-delta + 36 codec-rt-alp = 81
  workspace tests pass.
- 41 o11ytsdb-wasm tests pass (was 67; 26 ALP-area tests moved to
  the workspace crate).
- 561 TS tests pass repo-wide.
- typecheck + biome lint clean.
- Total Rust tests: 122 (up from 110 pre-M0).

Wasm size: 2,143,550 raw / 18,409 gz. Cumulative delta vs pre-M0
baseline: +210 raw (+0.010%) / +265 gz (+1.46%). Body extraction
added a tiny gz overhead (+20 bytes) from the new public function
boundaries — within noise.

PLAN.md updated: M0 marked complete.
